refactor: move provider profile features
