+++ src/lib/modules/providers/service/profiles/create-provider-profile-for-practice/transaction/createDirectoryListing.ts
@@ -1,4 +1,4 @@
-import { CreateProviderProfileForPractice } from '@/lib/modules/provider-profiles/features';
+import { CreateProviderProfileForPractice } from '@/lib/modules/providers/features/profiles';
 import { ListingStatus } from '@prisma/client';
 
 import { CreateProviderProfileForPracticeTransaction } from './definition';
